+++ NEWS.md
@@ -2,6 +2,7 @@
 
 User-visible changes worth mentioning.
 
+- [#823] Make configuration and specs ORM independent
 - [#777] Add support for public client in password grant flow
 
 ---